requirements.txt
crewai==0.30.11
langchain==0.2.1
langchain_community==0.2.1
langchain_google_genai==1.0.5
python-dotenv==1.0.1
streamlit==1.34.0
